Made Checkered work on Windows.  Removed Windows warnings in mrvImageBrowser and mrvImageInformation and mrvPreferences.
Former-commit-id: 9de9855ce83df444c5de6707d0055d1f96b8de16
Former-commit-id: e5176bb950da3d1668284815488d1dd3cde77279 [formerly ba8129b42fa8ead6f6c356e48ceef5b949aad5fe [formerly 9539944a80f2b4a4c469717139f83fbf5bd07cdd]]
Former-commit-id: ba8129b42fa8ead6f6c356e48ceef5b949aad5fe
Former-commit-id: cf96cf7cbbec5be8373b53f801ae65c2466ce240

+++ mrViewer/src/gui/mrvImageBrowser.h
@@ -261,7 +261,7 @@ protected:
     //! These functions send reel, images o current image to network
     void send_reel( const mrv::Reel& r );
     void send_images( const mrv::Reel& r);
-    void send_current_image( int idx, const mrv::media& m );
+    void send_current_image( size_t idx, const mrv::media& m );
 
     //! Change reel to new one selected
     void change_reel();
